chore: add manager for ddl builders
