+++ internal/exporter/factory/factory.go
@@ -36,6 +36,7 @@ func CreateExporters(renderer *template.Renderer) map[config.ExporterName]export
 	goModFinder := golang.NewModFinder()
 	goPropertyMapper := goentity.NewGoPropertyMapper()
 	graphBuilder := diagram.NewGraphBuilder(renderer)
+	ddlBuilderManager := sql.NewDDLBuilderManager()
 
 	return map[config.ExporterName]exporter.Exporter{
 		config.ExporterNameMd:      markdown.NewMarkdownExporter(pager, graphBuilder),
@@ -46,12 +47,12 @@ func CreateExporters(renderer *template.Renderer) map[config.ExporterName]export
 			goEntityGenerator,
 			goModFinder,
 		),
-		config.ExporterNameGoose: goose.NewMigrationsExporter(pager, sql.NewDDLBuilder()),
+		config.ExporterNameGoose: goose.NewMigrationsExporter(pager, ddlBuilderManager),
 		config.ExporterNameGoSQLMigrate: gosqlmigrate.NewSQLMigrateExporter(
 			pager,
-			sql.NewDDLBuilder(),
+			ddlBuilderManager,
 		),
-		config.ExporterNameLaravelMigrationsRaw: laravel.NewLaravelMigrationsRawExporter(pager, sql.NewDDLBuilder()),
+		config.ExporterNameLaravelMigrationsRaw: laravel.NewLaravelMigrationsRawExporter(pager, ddlBuilderManager),
 		config.ExporterNameLaravelModels:        laravel.NewLaravelModelsExporter(pager),
 		config.ExporterNameGrpcCrud:             grpccrud.NewCrudExporter(pager),
 		config.ExporterNameGooseFixtures:        goose.NewFixturesExporter(pager, dataLoader, sql.NewInsertBuilder()),
